Update ULP.h
"3) In the implementation of temperature compensation, be sure to pass the function a float value. If you pass an integer or a rounded float value, you will see large jumps in the concentration whenever the value changes, ie. from 24 C to 25 C."
But the function return is defined as INT ???

+++ ULP/src/ULP.h
@@ -33,7 +33,7 @@ class ULP {
 
     ULP (int a, int b, float c);
 
-    int getTemp(int n, char* U = "C");
+    float getTemp(int n, char* U = "C");
 
 	float getConc(int n, float t = 20);
